+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.14.2
+
+* satella futures are now generic
+